requirements.txt
ultralytics
opencv-python
numpy
torch
pillow
tk
torchvision
pyyaml